+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- add `envbar:status` Artisan command to display current status and metadata in console
+
 ## [0.1.2] - 2026-04-14
 
 ### Fixed